OpenGL: Split OpenGLBackend in Forward.h and DeviceGL
