+++ src/backend/opengl/BlendStateGL.cpp
@@ -14,7 +14,6 @@
 
 #include "backend/opengl/BlendStateGL.h"
 
-#include "backend/opengl/OpenGLBackend.h"
 #include "common/Assert.h"
 
 namespace backend { namespace opengl {
