+++ src/backend/opengl/DepthStencilStateGL.cpp
@@ -14,7 +14,6 @@
 
 #include "backend/opengl/DepthStencilStateGL.h"
 
-#include "backend/opengl/OpenGLBackend.h"
 #include "backend/opengl/PersistentPipelineStateGL.h"
 #include "common/Assert.h"
