+++ src/backend/opengl/SamplerGL.cpp
@@ -14,7 +14,7 @@
 
 #include "backend/opengl/SamplerGL.h"
 
-#include "backend/opengl/OpenGLBackend.h"
+#include "backend/opengl/DeviceGL.h"
 #include "common/Assert.h"
 
 namespace backend { namespace opengl {
